Split out create_run_settings
this functionality may be reused in other places
within the library, so the create_run_settings
function was placed inside smartsim/settings
and is now called from the experiment

Local tests have been adapted
