+++ smartsim/entity/ensemble.py
@@ -27,10 +27,9 @@
 from copy import deepcopy
 from os import getcwd
 
-from smartsim.error.errors import SmartSimError
 
-from ..error import EntityExistsError, SSUnsupportedError, UserStrategyError
-from ..settings.settings import BatchSettings, RunSettings
+from ..error import EntityExistsError, SSUnsupportedError, UserStrategyError, SmartSimError
+from ..settings.base import BatchSettings, RunSettings
 from ..utils import get_logger
 from ..utils.helpers import init_default
 from .entityList import EntityList
