+++ smartsim/experiment.py
@@ -31,7 +31,7 @@ from pprint import pformat
 
 import pandas as pd
 from tqdm import trange
-from . import settings
+from .settings import settings
 from .control import Controller, Manifest
 from .entity import Ensemble, Model
 from .error import SmartSimError
@@ -334,7 +334,6 @@ class Experiment:
                             run_command="auto",
                             run_args=None,
                             env_vars=None,
-                            skip_checks=False,
                             **kwargs):
         """Create a ``RunSettings`` instance.
 
@@ -342,6 +341,9 @@ class Experiment:
         match a run command on the system with a RunSettings
         class in SmartSim. If found, the class corresponding
         to that run_command will be created and returned.
+
+        if the local launcher is being used, auto detection will
+        be turned off.
 
         If a recognized run command is passed, the ``RunSettings``
         instance will be a child class such as ``SrunSetttings``
@@ -369,49 +371,14 @@ class Experiment:
         :return: the created ``RunSettings``
         :rtype: RunSettings
         """
-        # all supported RunSettings child classes
-        supported = {
-            "aprun": settings.AprunSettings,
-            "srun": settings.SrunSettings,
-            "mpirun": settings.MpirunSettings,
-            "jsrun": settings.JsrunSettings
-        }
-
-        # run commands supported by each launcher
-        by_launcher = {
-            "slurm": ["srun", "mpirun"], # TODO support arun for slurm?
-            "pbs": ["aprun", "mpirun"],
-            "cobalt": ["aprun", "mpirun"],
-            "lsf": ["jsrun", "mpirun"]
-        }
-
-        def _make_settings(cmd):
-            run_settings = supported[cmd](
-                    exe, exe_args, run_args, env_vars, **kwargs
-                )
-            return run_settings
-
         try:
-            # if local launcher we want to skip detection as a user
-            # may not want to run with a run_command at all
-            if run_command == "auto" and self._launcher != "local":
-                if self._launcher in by_launcher:
-                    for cmd in by_launcher[self._launcher]:
-                        if skip_checks or is_valid_cmd(cmd):
-                            return _make_settings(cmd)
-                raise SmartSimError(
-                    f"Could not automatically detect a run command to use for launcher {self._launcher}")
-
-            if run_command and run_command in supported:
-                return _make_settings(run_command)
-
-            # if not auto detecting, custom run_command, or running without run_command
-            if run_command == "auto":
-                run_command = None
-            run_settings = settings.RunSettings(
-                exe, exe_args, run_command, run_args, env_vars
-            )
-            return run_settings
+            return settings.create_run_settings(self._launcher,
+                                                exe,
+                                                exe_args=exe_args,
+                                                run_command=run_command,
+                                                run_args=run_args,
+                                                env_vars=env_vars,
+                                                **kwargs)
         except SmartSimError as e:
             logger.error(e)
             raise
