+++ smartsim/settings/__init__.py
@@ -3,5 +3,17 @@ from .cobaltSettings import CobaltBatchSettings
 from .lsfSettings import BsubBatchSettings, JsrunSettings
 from .mpirunSettings import MpirunSettings
 from .pbsSettings import QsubBatchSettings
-from .settings import RunSettings
+from .base import RunSettings
 from .slurmSettings import SbatchSettings, SrunSettings
+
+__all__ = [
+    "AprunSettings",
+    "CobaltBatchSettings",
+    "BsubBatchSettings",
+    "JsrunSettings",
+    "MpirunSettings",
+    "QsubBatchSettings",
+    "RunSettings",
+    "SbatchSettings",
+    "SrunSettings"
+]
