+++ tests/test_local_restart.py
@@ -13,7 +13,7 @@ def test_restart(fileutils):
     test_dir = fileutils.make_test_dir(exp_name)
 
     script = fileutils.get_test_conf_path("sleep.py")
-    settings = RunSettings("python", f"{script} --time=3")
+    settings = exp.create_run_settings("python", f"{script} --time=3")
 
     M1 = exp.create_model("m1", path=test_dir, run_settings=settings)
 
@@ -33,7 +33,7 @@ def test_ensemble(fileutils):
     test_dir = fileutils.make_test_dir(exp_name)
 
     script = fileutils.get_test_conf_path("sleep.py")
-    settings = RunSettings("python", f"{script} --time=3")
+    settings = exp.create_run_settings("python", f"{script} --time=3")
 
     ensemble = exp.create_ensemble("e1", run_settings=settings, replicas=2)
     ensemble.set_path(test_dir)
